refactor: added exceptions.ts file and utilities in snapUtils
