+++ packages/starknet-snap/src/signDeclareTransaction.ts
@@ -5,15 +5,8 @@ import {
   getKeysFromAddress,
   signDeclareTransaction as signDeclareTransactionUtil,
   validateAccountRequireUpgradeOrDeploy,
-  DeployRequiredError,
-  UpgradeRequiredError,
 } from './utils/starknetUtils';
-import {
-  getNetworkFromChainId,
-  getSignTxnTxt,
-  showDeployRequestModal,
-  showUpgradeRequestModal,
-} from './utils/snapUtils';
+import { getNetworkFromChainId, getSignTxnTxt, showAccountRequireUpgradeOrDeployModal } from './utils/snapUtils';
 import { heading, panel, DialogType } from '@metamask/snaps-sdk';
 import { logger } from './utils/logger';
 
@@ -28,12 +21,7 @@ export async function signDeclareTransaction(params: ApiParams): Promise<Signatu
     try {
       await validateAccountRequireUpgradeOrDeploy(network, signerAddress, publicKey);
     } catch (e) {
-      if (e instanceof DeployRequiredError) {
-        await showDeployRequestModal(wallet);
-      }
-      if (e instanceof UpgradeRequiredError) {
-        await showUpgradeRequestModal(wallet);
-      }
+      await showAccountRequireUpgradeOrDeployModal(wallet, e);
       throw e;
     }
