+++ packages/starknet-snap/src/signTransaction.ts
@@ -1,19 +1,8 @@
 import { toJson } from './utils/serializer';
 import { Signature } from 'starknet';
 import { ApiParams, SignTransactionRequestParams } from './types/snapApi';
-import {
-  getKeysFromAddress,
-  signTransactions,
-  validateAccountRequireUpgradeOrDeploy,
-  DeployRequiredError,
-  UpgradeRequiredError,
-} from './utils/starknetUtils';
-import {
-  getNetworkFromChainId,
-  getSignTxnTxt,
-  showDeployRequestModal,
-  showUpgradeRequestModal,
-} from './utils/snapUtils';
+import { getKeysFromAddress, signTransactions, validateAccountRequireUpgradeOrDeploy } from './utils/starknetUtils';
+import { getNetworkFromChainId, getSignTxnTxt, showAccountRequireUpgradeOrDeployModal } from './utils/snapUtils';
 import { heading, panel, DialogType } from '@metamask/snaps-sdk';
 import { logger } from '../src/utils/logger';
 
@@ -28,12 +17,7 @@ export async function signTransaction(params: ApiParams): Promise<Signature | bo
     try {
       await validateAccountRequireUpgradeOrDeploy(network, signerAddress, publicKey);
     } catch (e) {
-      if (e instanceof DeployRequiredError) {
-        await showDeployRequestModal(wallet);
-      }
-      if (e instanceof UpgradeRequiredError) {
-        await showUpgradeRequestModal(wallet);
-      }
+      await showAccountRequireUpgradeOrDeployModal(wallet, e);
       throw e;
     }
