+++ packages/starknet-snap/src/utils/exceptions.ts
@@ -1,0 +1,11 @@
+export class UpgradeRequiredError extends Error {
+  constructor(msg: string) {
+    super(msg);
+  }
+}
+
+export class DeployRequiredError extends Error {
+  constructor(msg: string) {
+    super(msg);
+  }
+}
