+++ packages/starknet-snap/test/src/signTransaction.test.ts
@@ -12,6 +12,7 @@ import { ApiParams, SignTransactionRequestParams } from '../../src/types/snapApi
 import { constants } from 'starknet';
 import * as utils from '../../src/utils/starknetUtils';
 import * as snapsUtil from '../../src/utils/snapUtils';
+import { DeployRequiredError, UpgradeRequiredError } from '../../src/utils/exceptions';
 
 chai.use(sinonChai);
 const sandbox = sinon.createSandbox();
@@ -65,6 +66,13 @@ describe('Test function: signTransaction', function () {
     sandbox.useFakeTimers(createAccountProxyTxn.timestamp);
     walletStub.rpcStubs.snap_dialog.resolves(true);
     walletStub.rpcStubs.snap_manageState.resolves(state);
+    sandbox.stub(snapsUtil, 'showAccountRequireUpgradeOrDeployModal').callsFake(async (wallet, e) => {
+      if (e instanceof DeployRequiredError) {
+        await snapsUtil.showDeployRequestModal(wallet);
+      } else if (e instanceof UpgradeRequiredError) {
+        await snapsUtil.showUpgradeRequestModal(wallet);
+      }
+    });
   });
 
   afterEach(function () {
@@ -83,7 +91,7 @@ describe('Test function: signTransaction', function () {
   it('should 1) throw an error and 2) show upgrade modal if account upgrade required', async function () {
     const validateAccountRequireUpgradeOrDeployStub = sandbox
       .stub(utils, 'validateAccountRequireUpgradeOrDeploy')
-      .throws(new utils.UpgradeRequiredError('Upgrade Required'));
+      .throws(new UpgradeRequiredError('Upgrade Required'));
     const showUpgradeRequestModalStub = sandbox.stub(snapsUtil, 'showUpgradeRequestModal').resolves();
     let result;
     try {
@@ -106,9 +114,7 @@ describe('Test function: signTransaction', function () {
     const validateAccountRequireUpgradeOrDeployStub = sandbox
       .stub(utils, 'validateAccountRequireUpgradeOrDeploy')
       .throws(
-        new utils.DeployRequiredError(
-          `Cairo 0 contract address ${account1.address} balance is not empty, deploy required`,
-        ),
+        new DeployRequiredError(`Cairo 0 contract address ${account1.address} balance is not empty, deploy required`),
       );
     const showDeployRequestModalStub = sandbox.stub(snapsUtil, 'showDeployRequestModal').resolves();
     let result;
